# Performance optimization of register_block_script_handle()