Optimize register_block_script_handle() to only call plugins_url() if core or theme block conditions are not met.

Trac ticket: https://core.trac.wordpress.org/ticket/59181