@mukeshpanchal27 2023-08-25
Committed in https://core.trac.wordpress.org/changeset/56455