@oandregal 2023-08-25
@mukeshpanchal27 @dougaxe1 would you be able to provide a before/after benchmarking? I mean, not a before/after profiling, but real numbers simulating a production environment – otherwise, we don't know the real impact.

If I understand how CI performance tests are setup for this (am I reading this right @felixarntz @joemcgill ?), this PR was committed as of https://github.com/WordPress/wordpress-develop/commit/0c82e15d72ee5114e61c676d6559d0763939d572 and so the performance tests action for it should be [this one](https://github.com/WordPress/wordpress-develop/actions/runs/5968148105/job/16191556893). According to those numbers, this has had a tremendous impact:

- TTFB for classic theme has gone from 66.15ms to 59.69ms: -6.46ms, 9.76% faster.
- TTFB for block themes has gone from 88.14ms to 62ms: -26ms, 29.65% faster.

| Before | After |
| --- | --- |
|  <img width="359" alt="Captura de ecrã 2023-08-25, às 12 56 19" src="https://github.com/WordPress/wordpress-develop/assets/583546/078eb545-a6a7-4a5c-bda2-52dfa42d1f9b"> | <img width="360" alt="Captura de ecrã 2023-08-25, às 12 57 13" src="https://github.com/WordPress/wordpress-develop/assets/583546/07765190-c26e-4a26-8633-6c6b9aa6eebd"> |