@dougaxe1 2023-08-25
Hey @oandregal, I too thought those numbers were  too good to be true. Luckily I was sitting next to @davidbumwald who explained that TTFB numbers fluctuate pretty wildly in these automated tests, and maybe aren't quite representative of actual impact. An XHProf benchmark would be more accurate for these micro-optimizations.

@mukeshpanchal27 did I get that right?