@oandregal 2023-08-25
> If I understand how CI performance tests are setup for this (am I reading this right @felixarntz @joemcgill ?)

Ah, I think the response here is that I misunderstood the "base results". According to [the workflow](https://github.com/WordPress/wordpress-develop/blob/trunk/.github/workflows/performance.yml#L166) it seems that it compares the actual commit with the "base", which is [6.1.1](https://github.com/WordPress/wordpress-develop/blob/trunk/.github/workflows/performance.yml#L45). We've gone a long way since then, so the difference makes sense now.