@oandregal 2023-08-25
Actually, I just found out about https://github.com/WordPress/wordpress-develop/pull/5000 which is all what I've been wishing for! :)