@dougaxe1 2023-08-25
Thank you for the follow-up. I clearly misunderstood the performance tests too. Thank you for explaining, and yes, hoping for traction on #5000!