@mukeshpanchal27 2023-08-28
@oandregal @dougaxe1 (cc. @joemcgill @felixarntz) I took a benchmarking for before after the change and it clear shows that it will improve Server side metric by ~1%. Check [Sheet](https://docs.google.com/spreadsheets/d/1PxW2bfXu5CpkcLVmRGS7fYWGs1XWjYQhy7KnNCRh0Fc/edit?usp=sharing).

  | Before | After | Diff % | Diff abs. |
-- | -- | -- | -- | -- |
wp-before-template (p10) | 174.37 | 173.14 | -0.71% | -1.23 |
wp-before-template (p25) | 176.22 | 175.01 | -0.69% | -1.21 |
wp-before-template (p50) | 179.74 | 177.84 | -1.06% | -1.90 |
wp-before-template (p75) | 187.68 | 180.18 | -4.00% | -7.50 |
wp-before-template (p90) | 197.96 | 182.97 | -7.57% | -14.99 |
wp-template (p10) | 194.4 | 192.85 | -0.80% | -1.55 |
wp-template (p25) | 195.73 | 194.32 | -0.72% | -1.41 |
wp-template (p50) | 198.42 | 195.89 | **-1.28%** | **-2.53** |
wp-template (p75) | 202.58 | 198.11 | -2.21% | -4.47 |
wp-template (p90) | 209.61 | 200.27 | -4.46% | -9.34 |
wp-total (p10) | 370.29 | 368.03 | -0.61% | -2.26 |
wp-total (p25) | 372.5 | 370.43 | -0.56% | -2.07 |
wp-total (p50) | 377.61 | 373.77 | -1.02% | -3.84 |
wp-total (p75) | 389.8 | 377.52 | -3.15% | -12.28 |
wp-total (p90) | 402.07 | 379.7 | -5.56% | -22.37 |